Adding custom-hook named as useDataFetch that include page numeration
